+++ README.md
@@ -50,7 +50,8 @@ module.exports = {
     // ... Your other themes.
     [
       require.resolve("@easyops-cn/docusaurus-search-local"),
-      {
+      /** @type {import("@easyops-cn/docusaurus-search-local").PluginOptions} */
+      ({
         // ... Your options.
         // `hashed` is recommended as long-term-cache of index file is possible.
         hashed: true,
@@ -58,7 +59,7 @@ module.exports = {
         // ```
         // language: ["en", "zh"],
         // ```
-      },
+      }),
     ],
   ],
 };
