+++ docusaurus-search-local/tsconfig.types.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "emitDeclarationOnly": true,
+    "declaration": true,
+    "declarationDir": "dist/types"
+  },
+  "include": ["src/index.ts"],
+}
