+++ website/docusaurus.config.js
@@ -119,12 +119,13 @@ module.exports = {
   themes: [
     [
       "@easyops-cn/docusaurus-search-local",
-      {
+      /** @type {import("@easyops-cn/docusaurus-search-local").PluginOptions} */
+      ({
         hashed: true,
         language: ["en", "zh"],
         highlightSearchTermsOnTargetPage: true,
         explicitSearchResultPath: true,
-      },
+      }),
     ],
   ],
 };
